Splitting RandomDownloadStrategy into two classes, consolidating 32/64-bit uncleanliness

+++ components/gnutella-core/src/main/java/com/limegroup/gnutella/downloader/Interval.java
@@ -14,13 +14,16 @@ public class Interval implements Serializable{
     public final int low;
     public final int high;
 
-    /** @requires low<=high */
-    public Interval(int low, int high) {
+    /** @requires low<=high 
+     * 
+     * Stub for making code 64-bit clean.
+     */
+    public Interval(long low, long high) {
         if(high < low)
             throw new IllegalArgumentException("low: " + low +
                                             ", high: " + high);
-        this.low=low;
-        this.high=high;
+        this.low=(int)low;
+        this.high=(int)high;
     }
     
     public Interval(int singleton) {
